Fix EffectiveRibInWriter.writeTable()
Overwrites should be treated very simply, by wiping all routes
and iterating over routes which are present.

This used to be possible in the BI implementation through
RIBSupport.routesPath(), but that was not carried over.

Fix this omission by defining proper accessor methods in RIBSupport
and using them to efficienly re-create eff-rib-in.

Change-Id: I1aea06ec4f082d11643566aed761fcc4092b6fcc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>