Fix RouteEntry.{add,remove}Route()
This API is interfacing with MD-SAL, where values are boxed,
hence these should be propagate as they are, not forcibly unboxed
in LocRibWriter.

This improves non-addpath performance, as we are simply skipping
the unboxing operation, since we do not care at all about pathIds.

Also hide RouteKey, as that is an implementation-internal class
which should not be exposed.

Change-Id: Ica7612ec5e6df8d07eb16294c1efa5fb0428f884
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>