Add adj-rib-in llgr-stale flag
Add the ability to signal tables to be LLGR-stale, so that
effective-rib-in can propagate this state properly towards
local rib.

This flag is picked up by before we are applying input policies,
so that routes are transiently marked as LLGR_STALE. The policy
can then decide whether to filter the route or modify its
attributes.

Once the policy has been applied and the flag is set, we check
if NO_LLGR flag is present, and if that is the case, we remove
that route.

This mechanics is required to correctly allow policies to
see LLGR_STALE routes as well as potentially remove the NO_LLGR
flag.

JIRA: BGPCEP-495
Change-Id: Iac9815b29e97aee99a81d015194cd16fb362387d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>