Optimize deletion of entire adj-rib-in table
In case we are deleting an entire table, there is no point in
issuing delete requests for its children.

This refactor eliminates the delete operations, invokes checks
for route target table only once and updates the counters only
once.

This is a stepping stone towards being able to handle LLGR_STALE
routes in an atomic fashion.

JIRA: BGPCEP-495
Change-Id: I3f345cd8764ad83bb39724fbfd42a0bcdc16a009
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>