Optimize whole-table overwrite
When a table is overwritten, we can treat it as a delete followed
by a fill of after-routes. This simplifies the logic quite a bit
and speeds it up. It also allows parts of it to operate on
before/after data rather than a DataTreeCandidateNode, hence it
is a reusable piece we'll need for LLGR_STALE nodes.

JIRA: BGPCEP-459
Change-Id: Ib84296652d84c4616d617583f14e5395058eac1c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>