Update patch set 2
Patch Set 2:

This is close the cleaning of whole PCEP project. Only 3 Errors/Warning remains for which I have no solution:

[WARNING] /home/ohyd6323/Soft/OpenDayLight/bgpcep/pcep/topology/topology-provider/src/test/java/org/opendaylight/bgpcep/pcep/topology/provider/AbstractPCEPSessionTest.java:[149,70] newInstance() in java.lang.Class has been deprecated
[ERROR] /home/ohyd6323/Soft/OpenDayLight/bgpcep/pcep/topology/topology-provider/src/main/java/org/opendaylight/bgpcep/pcep/topology/provider/config/PCEPTopologyProviderBean$PCEPTopologyProviderBeanCSS.java:143: Prefer java.util.HashMap
[ERROR] /home/ohyd6323/Soft/OpenDayLight/bgpcep/pcep/tunnel/tunnel-provider/src/main/java/org/opendaylight/bgpcep/pcep/tunnel/provider/PCEPTunnelClusterSingletonService.java:92: Prefer java.util.HashMap

Changing HashTable by HasMap is not compatible with Dictionary which could not be modified due to the dependence to OSGI class. And for the first one, I don't found the correct SuppressWarnings. If someone has an idea ...

Patch-set: 2
Label: Verified=0