Update patch set 12
Patch Set 12:

(2 comments)

> Patch Set 12: Code-Review-1
> 
> (3 comments)
> 
> From what I understand, the ietf-inet-types@2013-07-15 introduces zoned and no-zone IP addresses but the original internal YANG models shipped with ODL BGPCEP didn't use them at all.
> Even if this can be considered as a bug, the revision of YANG models should be changed anyway to avoid problems with previous versions released. Let me know if I can give an hand on this.

This is actually the most feasible compatibility-wise. Nothing changes about the endpoints, just zones which got accepted will not be (plus we guarantee we will never produce them). 99% of users are not even aware of the fact that inet:ipv4-address can contain a zone.

If we were to play this by the book, we would automatically break everyone interacting through RESTCONF (JSON: different module name, XML: different namespace). So we would need a compatibility layer and ship both modules.

I strongly believe going through the hoops has near-zero real value here.

> Also, I would have been split this changes into smaller pieces to migrate each artifact separately but I am not sure if this is possible without breaking the test suite.

This is a mechanic change involving common utilities, so it is not done easily. I do not see much value in trying to break it up.

Patch-set: 12
Label: Verified=0