Update patch set 1
Patch Set 1:

Hello,

I need to solve an issue when adding the features inside Karaf. But, I prefer to push the change now to let you more time to review it. Samcra.java is the hard part to read due to the complexity of the algorithm. If you have time, start reading the article from which the algorithm is issued for better understanding.

Olivier

Patch-set: 1
Label: Verified=0