Update patch set 1
Patch Set 1:

(19 comments)

Made all requested corrections and add 2 new Jira tickets: one for the SAMCRA bug and one for the documentation.

Unfortunately, I made the correction on the stable/magnesium change and then cherry pick on master without changing the Change-Id. So, now we have 2 similar changes against the master branch. But, the good new is that the new one is mark as cherry pick of the stable/magnesium branch.

Patch-set: 1