Update patch set 2
Patch Set 2:

Robert, 

The problem has been documented in the docs/pcep/pcep-path-computation.rst file as well as copy paste in the JIRA ticket.

What I could observe when it happens, is that the bgp-topology-provider bundle is placed in 'GracePeriod' state waiting the GraphProviderService to be up. The problem comes when Karaf detects that the Graph Service is up and provide a Null pointer instead the real pointer for this Graph Service. Thus, the LinkstateGraphBuilder() class constructor get a null pointer instead the link to the Graph Service. In parallel, the Path Computation Server which also needs the Graph Service is correctly launch as it is started after the Graph Service. Another work around would be to start the Gaph bundle before the bgp-topology-provider bundle.

Hope it provides a sufficient description of the problem, which, I persist, is a Karaf bug, not a BGPCEP bug.

Patch-set: 2