Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Robert, 
> 
> The problem has been documented in the docs/pcep/pcep-path-computation.rst file as well as copy paste in the JIRA ticket.
> 
> What I could observe when it happens, is that the bgp-topology-provider bundle is placed in 'GracePeriod' state waiting the GraphProviderService to be up. The problem comes when Karaf detects that the Graph Service is up and provide a Null pointer instead the real pointer for this Graph Service. Thus, the LinkstateGraphBuilder() class constructor get a null pointer instead the link to the Graph Service. In parallel, the Path Computation Server which also needs the Graph Service is correctly launch as it is started after the Graph Service. Another work around would be to start the Gaph bundle before the bgp-topology-provider bundle.
> 
> Hope it provides a sufficient description of the problem, which, I persist, is a Karaf bug, not a BGPCEP bug.

Olivier, I think we have a disconnect in processes. We are an open-source project collaborating with others. I do not question your claim this would somehow be a Blueprint/Karaf issue.

What I would like to point out is:
- pointing out that there is a bug without filing it with upstream is wrong, because *if* it is an upstream bug, they will never know about it
- even if we do not file an upstream issue (because we are not sure that it is an upstream issue), we need to have evidence of investigation -- things like karaf.log, with some amount of analysis
- I do not remember this happening before, so *something* is going on

If we do not have this tracked (and pointed to by a FIXME: introduced in this patch), this issue will *never* get fixed and 3 years down the road people will wonder.

I mean, from engineering perspective, if this is a random thing, have you analyzed logs from a got and bad runs to see if sequencing of events changed?

This is supposed to be production code, so it is very hard for me to shrug this off -- there is a perspective where reverting the entire integration is a better option...

Patch-set: 2
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>