Enforce field finality
We really want to safely publish whatever we have gotten and want
to reject nulls. This does not solve the underlying issue, but takes
a step in bracketing where it lies.

Adding LOG seems improve thread synchronisation. With that modification
the problem seems not reproductible in particular when karaf is restarted.

JIRA: BGPCEP-898
Change-Id: Id465a5dce3c0e4aaeefbaf1dc6f7d36a2a02712b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>