Update patch set 3
Patch Set 3:

Robert, I let revert the change. It is just to highlight the race condition which disappear when some LOG are added as this change the thread synchronisation.

Of course it is perfectible, and I'm pretty sure that there is other way to synchronised / schedule threads. I just discover that when I add some LOGS to trace what's happen, and in particular if the ConnectedGraphProvider is publish at the same address that consumers get the service, I can't reproduce the problem, in particular after restarting karaf while it always failed previously.

Patch-set: 3
Reviewer: Gerrit User 7231 <7231@7fc14799-209e-464c-9743-7a06c2c21a81>