Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Robert, I let revert the change. It is just to highlight the race condition which disappear when some LOG are added as this change the thread synchronisation.
> 
> Of course it is perfectible, and I'm pretty sure that there is other way to synchronised / schedule threads. I just discover that when I add some LOGS to trace what's happen, and in particular if the ConnectedGraphProvider is publish at the same address that consumers get the service, I can't reproduce the problem, in particular after restarting karaf while it always failed previously.

well, depending on the underlying environment the LOG may just be helping in one place but we'd
still hit the race in another.

BUT, did you see this? https://git.opendaylight.org/gerrit/c/bgpcep/+/88414
I think that is the work on top of this patch to get the real fix in place.

Patch-set: 3