Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Tejas, Olivier, given that you +1'd this patch, can you shed some light on the questions raised?

Robert, this issue started with a deadlock, in PCEPSessionImpl. 
Once that was fixed, the next deadlock occurred in TopologyStatsProvider.
Then the next one was in AbstractTopologySessionListener.

So given that it took fixing in all these areas to avoid the deadlock altogether, i feel we can retain a single change. If we consider fixing each of these issues in different gerrits, they may be incomplete from the perspective of avoiding the deadlock in this scenario.

The test is to stop and start a remote pcep session in a short duration.

Patch-set: 2
Reviewer: Gerrit User 8351 <8351@7fc14799-209e-464c-9743-7a06c2c21a81>