Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Tejas, Olivier, given that you +1'd this patch, can you shed some light on the questions raised?
> 
> Robert, this issue started with a deadlock, in PCEPSessionImpl. 
> Once that was fixed, the next deadlock occurred in TopologyStatsProvider.
> Then the next one was in AbstractTopologySessionListener.
> 
> So given that it took fixing in all these areas to avoid the deadlock altogether, i feel we can retain a single change. If we consider fixing each of these issues in different gerrits, they may be incomplete from the perspective of avoiding the deadlock in this scenario.
> 
> The test is to stop and start a remote pcep session in a short duration.

I disagree. If it is three different deadlocks, each of them has to have a self-contained change, so that it can actually be reviewed for correctness. If it takes three patches to completely cover the test case, which btw. is not covered by any UT here or anything, then so be it.

Review of threading interactions is *critical* and needs to be thorough -- see how that drop in close() 'synchronized' removes a synchronization point with dead timer. That is a real regression, unless you can show its superfluous.

Guys, this is complex code, and wave-through reviews are a sure way to arrive at a point where it will be a) fragile b) noone will know what's being done *why*.

Just imagine yourself finding that this patch broke something 5 years down the line. How much evidence have you left for your future self? I can guarantee that you will not remember any of the reasoning that feels obvious right now.

Patch-set: 2
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>