Update patch set 8
Patch Set 8:

(1 comment)

> Patch Set 8:
> 
> Sir, I tested payload on ODL and got desired output.
> A bit of issues were trivial and were removed.
> However, I got 400 Bad Request on one of the payload.
> Kindly, take a look at line 22
> 
> Build was a failure.
> however, the build log https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/bgpcep-rtd-verify-aluminium/82/html/bgp/bgp-user-guide-protocol-configuration.html showed correct output.
> As you suggested earlier, I guess build failure is due to version migration.
> Regards
> -Chinmay

The REST error is because when using POST, the body should not include the node specified in the URL (e.g. 'protocols'), for this reason the last posted payload should work fine with just 'protocol'. The build failure is because of the ODL version migration.

Patch-set: 8
Reviewer: Gerrit User 797 <797@7fc14799-209e-464c-9743-7a06c2c21a81>