WIP: Convert config-loader to OSGi DS
The primary task here is getting rid of explicit blueprint while
retaining autowiring functionality.

This forces us to decompose ConfigLoaderImpl into multiple classes,
which ends up being a boon for tests, as they can completely forgo
threading.

First of all FileWatcherImpl is renamed to DefaultFileWatcher and
is a completely separate component, with proper lifecycle without
edge case failures.

Next we split out AbstractConfigLoader with all the loading logic
and AbstractWatchingConfigLoader with the addition of having notion
of a FileWatcher. This allows us to use synchronous loading in
test harness, while also providing the ability to have async
processing.

We then instantiate SimpleConfigLoader and OSGiConfigLoader to
handle the cases of static vs. dynamic codecs, as well as
TestConfigLoader to handle unit testing tasks.

// FIXME: add explicit watcher tests

JIRA: BGPCEP-905
Change-Id: I5549bdfeb372b2f9cb15a1690c399401432f4702
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>