precise OpenConfig models original license
add a NOTICE and an openconfig.LICENSE files in the bgp module

OpenConfig YANG models are published under the Apache License version 2
a.k.a. ALv2 but they do not precise it in their description.

Though Alv2 allows redistribution under other licenses, it includes a
redistribution clause (4) that instructs to give any other recipients
of the work or the derivative works a copy of the original license.

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Id64f9473ebd43df0c851a0c78533ea593634fe60