Improve config loader performance
Gloval context does not change often and hence to is appropriate
to find the root schema node we need for parsing when that happens
and cache it.

This also means we can perform upfront checks and neatly ignore
any registrations which are not safisfied by the model context.

Furthermore OSGiConfigLoader is updated to take mode updates without
restarting, improving overall experience.

Since we are in the area, we also harden XML parsing through use
of UntrustedXML.

Change-Id: If9b17686e648294a81249ae1b40684f1448380fb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>