Remove RIBSupport.emptyRoutesContainer()
This method is not called anywhere simply because this container
has no semantics in the datastore. Remove the method and all its
implementations.

This means that extensions do not load the associated Builder and
its implementation class. The second is more important, as it means
that the routes container will have only one implementation (from
mdsal-dom-codec) -- allowing JIT to devirtualuze based on CHA.

Change-Id: I643185ee9f6d035235e6364b5832b6b128443bd3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>