Rework ribOut interaction
As it turns out, we are using a binding chain separately from the
main chain. The reason for this is that we use domChain for processing
towards from the peer -- which is common for both. bindingChain is
used exclusively for ribOut writeout.

Rename it and correct the guard/use mismatch we introduced in the
previous patch.

JIRA: BGPCEP-853
Change-Id: Idac28712ae9603f7cbfefe613baf2b47c7daff17
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>