Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> I believe this should be handled by ietf-stateful augmenting instantiations

It is how I proceed at the beginning. Just have a look to the patch set 1. However, this doesn't compile correctly. More exactly, maven was not able to produce the jar file due to cyclic dependency. Indeed, pcep-base-parser need pcep-ietf-stateful for the definition of the LSP Object and pcep-ietf-sateful need pcep-base-parser for the common object. Thus maven doesn't how resolve this chicken and egg problem. Of course, if you could solve this compilation issue, I agree that using augmentation is much more simpler.

Patch-set: 5