WIP: Add PCEPTopoloy{Singleton,Tracker}
Rework topology instantiation to split the individual concerns.
PCEPTopologyTracker is responsible for finding out when a particular
topology becomes interesting to the underlying code (and when becomes
uninteresting).

For each such topology, a PCEPTopologySingleton is instantiated. It then
handles allocation of the underlying InstructionScheduler as well as
bootstrapping a cluster-wide singleton PCEPTopologyProvider instance.

// FIXME: now PCEPTopologyProvider needs to track configuration and
//        update the dispatcher, etc.

// FIXME: switch blueprint over to PCEPTopologyTracker
// FIXME: remove the entire .config package

JIRA: BGPCEP-893
Change-Id: I56c47b89ab87cf33a6be1016fa132b6c1aa98cbb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>