Add Bandwidth Reservation in PCE server
With RSVP-TE, bandwidth on links belonging to the LSP are automatically
adjusted by the routers and re-advertised by the IGP.
However, due to lack of signalling in Segment Routing this bandwidth
adjustement does not occur.
Thus, the PCE server should manage by itself the reserved bandwidth of
Segment Routing path. In addition, routers do not behave the
same way regarding UnReservedBandwidth attributes.

This patch provides an internal mechanism to manage Bandwidth
Reservation to the PCE Server. It concerns 4 main parts:
- ConnectedEdge: add new Reserved Bandwidth attributes with
  dedicated method to get, add and remove bandwidth
- AbstractPathComputation: Adjust pruneEdge method to take
  care of this new Reserved Bandwidth attributes
- ManagedTePath: methods to add, update and delete bandwidth
  of the LSP on edges belonging to the computed path
- PathManagerProvider: Manage reserved bandwidth by calling
  new ManagedTePath methods when needed.

JIRA: BGPCEP-979
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: I45534f43478a4605165fbe4b9190cf3feccd6380