Add Close loop mechanism to PCE Server
To ease the management of TE Path, this patch adds a new mechanism
to the Connected Graph in order to automatically trigger actions
when an Edge or Vertex is added, updated or removed.

This allows the PCE server to register triggers on Edges and Vertices
which belong to a Managed TE Path. In case of problem, failure or
TE attributes modification, the PCE server could detect them and apply
correction e.g. re-route the path if new delay on Edge exceeded the
constraints.

JIRA: BGPCEP-979

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: I3071d324fc8c678da439258bee79202fa6c39267