Clear node state before releasing it
ServerSessionManager.releaseNodeState() issues a datastore operation
which deletes the entire node. This can complete quickly enough that
the follow-up call to unbind()'s attempt to delete the node already
sees it deleted -- and can thus fail with IAE as per datastore rules.

Fix this by first clearing the state and then releasing it.

JIRA: BGPCEP-1005
Change-Id: I65bcee1d8cb756e774ce0ecd6df92bcf743cd06c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>