Correct ReachTlvParser to handle IPv6 prefix
ReachTlvParser class support Ipv6 prefix only in serialization
methods. ParseTlvBody is only deal with IPv4 prefix. Thus, if
a peer BGP Link State router sends an Update Message with an
IPv6 prefix in IP Reachability Information NLRI, ParseTlvBody
method failed to parse the IPv6 prefix and assert a
BGPDocumentedException.

This patch looks to the length of the buffer to be parsed to
determine if the prefix is an IPv4 or an IPv6 one. SerializeModel
method has been also updated to avoid exception to determine
if the prefix is an IPv4 or an IPv6 one.

JIRA: BGPCEP-1009

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Change-Id: Ia5bac0fb2a6087e5c5cfbca0e3eee7d8fd7b822c